requirements.txt
Flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
werkzeug==3.0.3
elevenlabs==1.5.0
